requirements.txt
requests
flask
python-dotenv
supabase
eth-account
x402
httpx
asyncio